+++ cmd/wharf/provisioner.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"github.com/iver-wharf/wharf-cmd/pkg/provisioner"
 	"github.com/spf13/cobra"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
@@ -11,7 +12,18 @@ var provisionerFlags = struct {
 
 	restConfig *rest.Config
 	namespace  string
-}{}
+
+	instanceID string
+}{
+	instanceID: "local",
+}
+
+func newProvisioner() (provisioner.Provisioner, error) {
+	return provisioner.NewK8sProvisioner(
+		provisionerFlags.instanceID,
+		rootConfig.Provisioner,
+		provisionerFlags.restConfig)
+}
 
 var provisionerCmd = &cobra.Command{
 	Use:   "provisioner",
@@ -45,5 +57,6 @@ orchestration is handled inside the Kubernetes cluster, in comparison to the
 func init() {
 	rootCmd.AddCommand(provisionerCmd)
 
+	provisionerCmd.Flags().StringVar(&provisionerFlags.instanceID, "instance", provisionerFlags.instanceID, "Wharf instance ID, used to avoid collisions in Pod ownership.")
 	addKubernetesFlags(provisionerCmd.PersistentFlags(), &provisionerFlags.k8sOverrides)
 }