+++ pkg/provisioner/k8sprovisioner.go
@@ -2,9 +2,13 @@ package provisioner
 
 import (
 	"context"
+	"errors"
 	"fmt"
+	"path/filepath"
+	"strconv"
 
 	"github.com/iver-wharf/wharf-cmd/pkg/config"
+	"gopkg.in/typ.v3"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
@@ -12,22 +16,18 @@ import (
 	"k8s.io/client-go/rest"
 )
 
-var listOptionsMatchLabels = metav1.ListOptions{
-	LabelSelector: "app.kubernetes.io/name=wharf-cmd-worker," +
-		"app.kubernetes.io/managed-by=wharf-cmd-provisioner," +
-		"wharf.iver.com/instance=prod",
-}
-
 type k8sProvisioner struct {
-	Config     config.ProvisionerConfig
-	Clientset  *kubernetes.Clientset
-	Pods       corev1.PodInterface
-	restConfig *rest.Config
+	Config                 config.ProvisionerConfig
+	Clientset              *kubernetes.Clientset
+	Pods                   corev1.PodInterface
+	restConfig             *rest.Config
+	instanceID             string
+	listOptionsMatchLabels metav1.ListOptions
 }
 
 // NewK8sProvisioner returns a new Provisioner implementation that targets
 // Kubernetes using a specific Kubernetes namespace and REST config.
-func NewK8sProvisioner(config config.ProvisionerConfig, restConfig *rest.Config) (Provisioner, error) {
+func NewK8sProvisioner(instanceID string, config config.ProvisionerConfig, restConfig *rest.Config) (Provisioner, error) {
 	clientset, err := kubernetes.NewForConfig(restConfig)
 	if err != nil {
 		return nil, err
@@ -37,11 +37,17 @@ func NewK8sProvisioner(config config.ProvisionerConfig, restConfig *rest.Config)
 		Clientset:  clientset,
 		Pods:       clientset.CoreV1().Pods(config.K8s.Namespace),
 		restConfig: restConfig,
+		instanceID: instanceID,
+		listOptionsMatchLabels: metav1.ListOptions{
+			LabelSelector: "app.kubernetes.io/name=wharf-cmd-worker," +
+				"app.kubernetes.io/managed-by=wharf-cmd-provisioner," +
+				"wharf.iver.com/instance=" + instanceID,
+		},
 	}, nil
 }
 
 func (p k8sProvisioner) ListWorkers(ctx context.Context) ([]Worker, error) {
-	podList, err := p.listPods(ctx, listOptionsMatchLabels)
+	podList, err := p.listPods(ctx, p.listOptionsMatchLabels)
 	if err != nil {
 		return []Worker{}, err
 	}
@@ -62,14 +68,17 @@ func (p k8sProvisioner) DeleteWorker(ctx context.Context, workerID string) error
 	return p.Pods.Delete(ctx, pod.Name, metav1.DeleteOptions{})
 }
 
-func (p k8sProvisioner) CreateWorker(ctx context.Context) (Worker, error) {
-	podMeta := createPodMeta(p.Config.Worker)
+func (p k8sProvisioner) CreateWorker(ctx context.Context, args WorkerArgs) (Worker, error) {
+	if args.GitCloneURL == "" {
+		return Worker{}, errors.New("missing required Git clone URL")
+	}
+	podMeta := p.newWorkerPod(args)
 	newPod, err := p.Pods.Create(ctx, &podMeta, metav1.CreateOptions{})
 	return convertPodToWorker(newPod), err
 }
 
 func (p k8sProvisioner) getPod(ctx context.Context, workerID string) (*v1.Pod, error) {
-	podList, err := p.listPods(ctx, listOptionsMatchLabels)
+	podList, err := p.listPods(ctx, p.listOptionsMatchLabels)
 	if err != nil {
 		return nil, err
 	}
@@ -81,7 +90,7 @@ func (p k8sProvisioner) getPod(ctx context.Context, workerID string) (*v1.Pod, e
 	return nil, fmt.Errorf("found no worker with appropriate labels matching workerID: %s", workerID)
 }
 
-func createPodMeta(config config.WorkerPodConfig) v1.Pod {
+func (p k8sProvisioner) newWorkerPod(args WorkerArgs) v1.Pod {
 	const (
 		repoVolumeName      = "repo"
 		repoVolumeMountPath = "/mnt/repo"
@@ -92,9 +101,9 @@ func createPodMeta(config config.WorkerPodConfig) v1.Pod {
 		"app.kubernetes.io/part-of":    "wharf",
 		"app.kubernetes.io/managed-by": "wharf-cmd-provisioner",
 		"app.kubernetes.io/created-by": "wharf-cmd-provisioner",
-		"wharf.iver.com/instance":      "prod",
-		"wharf.iver.com/build-ref":     "123",
-		"wharf.iver.com/project-id":    "456",
+		"wharf.iver.com/instance":      typ.Coal(args.WharfInstanceID, p.instanceID),
+		"wharf.iver.com/build-ref":     uitoa(args.BuildID),
+		"wharf.iver.com/project-id":    uitoa(args.ProjectID),
 	}
 	volumeMounts := []v1.VolumeMount{
 		{
@@ -103,84 +112,89 @@ func createPodMeta(config config.WorkerPodConfig) v1.Pod {
 		},
 	}
 
+	gitArgs := []string{"git", "clone", args.GitCloneURL, "--single-branch"}
+	if args.GitCloneBranch != "" {
+		gitArgs = append(gitArgs, "--branch", args.GitCloneBranch)
+	}
+	gitArgs = append(gitArgs, repoVolumeMountPath)
+
+	wharfArgs := []string{"run", "--loglevel", "debug", "--serve"}
+	if args.Environment != "" {
+		wharfArgs = append(wharfArgs, "--environment", args.Environment)
+	}
+	if args.Stage != "" {
+		wharfArgs = append(wharfArgs, "--stage", args.Stage)
+	}
+	for k, v := range args.Inputs {
+		wharfArgs = append(wharfArgs, "--input", fmt.Sprintf("%s=%s", k, v))
+	}
+	if args.SubDir != "" {
+		relSubDir := filepath.ToSlash(args.SubDir)
+		if filepath.IsAbs(relSubDir) {
+			relSubDir = "." + relSubDir
+		}
+		// Need "--" so the path isn't malliciously treated as another flag
+		wharfArgs = append(wharfArgs, "--", relSubDir)
+	}
+
+	wharfEnvs := []v1.EnvVar{
+		{
+			Name:  "WHARF_KUBERNETES_OWNER_ENABLE",
+			Value: "true",
+		},
+		{
+			Name: "WHARF_KUBERNETES_OWNER_NAME",
+			ValueFrom: &v1.EnvVarSource{
+				FieldRef: &v1.ObjectFieldSelector{
+					APIVersion: "v1",
+					FieldPath:  "metadata.name",
+				},
+			},
+		},
+		{
+			Name: "WHARF_KUBERNETES_OWNER_UID",
+			ValueFrom: &v1.EnvVarSource{
+				FieldRef: &v1.ObjectFieldSelector{
+					APIVersion: "v1",
+					FieldPath:  "metadata.uid",
+				},
+			},
+		},
+	}
+
+	for k, v := range args.AdditionalVars {
+		wharfEnvs = append(wharfEnvs, v1.EnvVar{
+			Name:  "WHARF_VAR_" + k,
+			Value: stringify(v),
+		})
+	}
+
 	return v1.Pod{
 		ObjectMeta: metav1.ObjectMeta{
 			GenerateName: "wharf-cmd-worker-",
 			Labels:       labels,
 		},
 		Spec: v1.PodSpec{
-			ServiceAccountName: config.ServiceAccountName,
+			ServiceAccountName: p.Config.Worker.ServiceAccountName,
 			RestartPolicy:      v1.RestartPolicyNever,
 			InitContainers: []v1.Container{
 				{
 					Name:            "init",
-					Image:           fmt.Sprintf("%s:%s", config.InitContainer.Image, config.InitContainer.ImageTag),
-					ImagePullPolicy: config.InitContainer.ImagePullPolicy,
-					// TODO: Should use repo URL and branch from build params.
-					Args: []string{
-						"git",
-						"clone",
-						"--single-branch",
-						"--branch", "feature/set-k8s-metadata-on-build-pods",
-						"https://github.com/iver-wharf/wharf-cmd",
-						repoVolumeMountPath,
-					},
-					VolumeMounts: volumeMounts,
+					Image:           fmt.Sprintf("%s:%s", p.Config.Worker.InitContainer.Image, p.Config.Worker.InitContainer.ImageTag),
+					ImagePullPolicy: p.Config.Worker.InitContainer.ImagePullPolicy,
+					Command:         gitArgs,
+					VolumeMounts:    volumeMounts,
 				},
 			},
 			Containers: []v1.Container{
 				{
-					Name: "app",
-					// TODO: Do some research on which image would be best to use.
-					//
-					// Note: golang:latest was faster than ubuntu:20.04 up until
-					// running `go install`, by around 20 seconds.
-					//
-					// Note2: The testing environment's internet speed was very
-					// fast, so the larger size:
-					//  Go: 353MB compressed
-					//  ubuntu: 73 MB uncompressed
-					// was barely a factor.
-					Image:           fmt.Sprintf("%s:%s", config.Container.Image, config.Container.ImageTag),
-					ImagePullPolicy: config.Container.ImagePullPolicy,
-					// TODO: Needs better implementation.
-					// Currently works for wharf-cmd, but is not thought through. Takes a long
-					// time to get running.
-					Command: []string{
-						"/bin/sh", "-c",
-						`
-make deps-go swag install && \
-cd test && \
-export WHARF_VAR_CHART_REPO="chart_repo" && \
-export WHARF_VAR_REG_URL="reg_url" && \
-wharf run --serve --stage test --loglevel debug`,
-					},
-					WorkingDir:   repoVolumeMountPath,
-					VolumeMounts: volumeMounts,
-					Env: []v1.EnvVar{
-						{
-							Name:  "WHARF_KUBERNETES_OWNER_ENABLE",
-							Value: "true",
-						},
-						{
-							Name: "WHARF_KUBERNETES_OWNER_NAME",
-							ValueFrom: &v1.EnvVarSource{
-								FieldRef: &v1.ObjectFieldSelector{
-									APIVersion: "v1",
-									FieldPath:  "metadata.name",
-								},
-							},
-						},
-						{
-							Name: "WHARF_KUBERNETES_OWNER_UID",
-							ValueFrom: &v1.EnvVarSource{
-								FieldRef: &v1.ObjectFieldSelector{
-									APIVersion: "v1",
-									FieldPath:  "metadata.uid",
-								},
-							},
-						},
-					},
+					Name:            "app",
+					Image:           fmt.Sprintf("%s:%s", p.Config.Worker.Container.Image, p.Config.Worker.Container.ImageTag),
+					ImagePullPolicy: p.Config.Worker.Container.ImagePullPolicy,
+					Args:            wharfArgs,
+					WorkingDir:      repoVolumeMountPath,
+					VolumeMounts:    volumeMounts,
+					Env:             wharfEnvs,
 				},
 			},
 			Volumes: []v1.Volume{
@@ -192,5 +206,20 @@ wharf run --serve --stage test --loglevel debug`,
 				},
 			},
 		},
+	}
+}
+
+func uitoa(i uint) string {
+	return strconv.FormatUint(uint64(i), 10)
+}
+
+func stringify(value any) string {
+	switch value := value.(type) {
+	case string:
+		return value
+	case nil:
+		return "" // fmt.Sprint returns "<nil>", which we don't want
+	default:
+		return fmt.Sprint(value)
 	}
 }