+++ pkg/provisionerapi/provisionerapi.go
@@ -4,14 +4,12 @@ import (
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
 	"github.com/iver-wharf/wharf-cmd/pkg/config"
+	"github.com/iver-wharf/wharf-cmd/pkg/provisioner"
+	"github.com/iver-wharf/wharf-cmd/pkg/provisionerapi/docs"
 	"github.com/iver-wharf/wharf-core/pkg/ginutil"
 	"github.com/iver-wharf/wharf-core/pkg/logger"
 	ginSwagger "github.com/swaggo/gin-swagger"
 	"github.com/swaggo/gin-swagger/swaggerFiles"
-	"k8s.io/client-go/rest"
-
-	// Load in swagger docs
-	_ "github.com/iver-wharf/wharf-cmd/pkg/provisionerapi/docs"
 )
 
 var log = logger.NewScoped("PROVISIONER-API")
@@ -25,9 +23,8 @@ var log = logger.NewScoped("PROVISIONER-API")
 // @contact.name Iver wharf-cmd support
 // @contact.url https://github.com/iver-wharf/wharf-cmd/issues
 // @contact.email wharf@iver.se
-// @basePath /api
 // @query.collection.format multi
-func Serve(config config.Config, cfg *rest.Config) error {
+func Serve(prov provisioner.Provisioner, config config.ProvisionerAPIConfig) error {
 	gin.DefaultWriter = ginutil.DefaultLoggerWriter
 	gin.DefaultErrorWriter = ginutil.DefaultLoggerWriter
 
@@ -37,26 +34,21 @@ func Serve(config config.Config, cfg *rest.Config) error {
 		ginutil.RecoverProblem,
 	)
 
-	applyCORS(r, config.ProvisionerAPI.HTTP.CORS)
+	applyCORS(r, config.HTTP.CORS)
 
-	g := r.Group("/api")
-	g.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
-	g.GET("", pingHandler)
+	r.GET("", pingHandler)
+	api := r.Group("/api")
+	api.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler, func(c *ginSwagger.Config) {
+		c.InstanceName = docs.SwaggerInfoprovisionerapi.InstanceName()
+	}))
 
-	workerModule := workerModule{}
-	if err := workerModule.init(config.Provisioner, cfg); err != nil {
+	workerModule{prov: prov}.register(api)
+
+	log.Info().WithString("address", config.HTTP.BindAddress).Message("Starting server.")
+	if err := r.Run(config.HTTP.BindAddress); err != nil {
 		log.Error().
 			WithError(err).
-			Message("Failed to initialize worker module.")
-		return err
-	}
-	workerModule.register(g)
-
-	log.Info().WithString("address", config.ProvisionerAPI.HTTP.BindAddress).Message("Starting server.")
-	if err := r.Run(config.ProvisionerAPI.HTTP.BindAddress); err != nil {
-		log.Error().
-			WithError(err).
-			WithString("address", config.ProvisionerAPI.HTTP.BindAddress).
+			WithString("address", config.HTTP.BindAddress).
 			Message("Failed to start web server.")
 		return err
 	}
@@ -93,6 +85,7 @@ type Ping struct {
 // @description Pong.
 // @description Added in v0.8.0.
 // @tags meta
+// @produce json
 // @success 200 {object} Ping
 // @router / [get]
 func pingHandler(c *gin.Context) {