+++ pkg/workerapi/workerserver/artifact.go
@@ -20,11 +20,23 @@ type ArtifactFileOpener interface {
 	OpenArtifactFile(artifactID uint) (io.ReadCloser, error)
 }
 
-func (m *artifactModule) register(g *gin.RouterGroup) {
+func (m artifactModule) register(g *gin.RouterGroup) {
 	g.GET("/artifact/:artifactId/download", m.downloadArtifactHandler)
 }
 
-func (m *artifactModule) downloadArtifactHandler(c *gin.Context) {
+// downloadArtifactHandler godoc
+// @id downloadArtifact
+// @summary Download an artifact file.
+// @description Added in v0.8.0.
+// @tags worker
+// @produce octet-stream
+// @param artifactId path uint true "Artifact ID" minimum(0)
+// @success 200 {file} string "OK"
+// @failure 400 {object} problem.Response "Bad request"
+// @failure 404 {object} problem.Response "Cannot find artifact"
+// @failure 502 {object} problem.Response "Canont read artifact file"
+// @router /api/artifact/{artifactId}/download [post]
+func (m artifactModule) downloadArtifactHandler(c *gin.Context) {
 	artifactID, ok := ginutil.ParseParamUint(c, "artifactId")
 	if !ok {
 		return