+++ pkg/workerapi/workerserver/restserver.go
@@ -5,12 +5,11 @@ import (
 
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
+	"github.com/iver-wharf/wharf-cmd/pkg/workerapi/workerserver/docs"
 	"github.com/iver-wharf/wharf-core/pkg/ginutil"
+	ginSwagger "github.com/swaggo/gin-swagger"
+	"github.com/swaggo/gin-swagger/swaggerFiles"
 )
-
-type module interface {
-	register(g *gin.RouterGroup)
-}
 
 type restServer struct {
 	artifactOpener ArtifactFileOpener
@@ -22,25 +21,29 @@ func newRestServer(artifactOpener ArtifactFileOpener) *restServer {
 	}
 }
 
+// serveHTTP godoc
+// @title Wharf worker API
+// @description REST API for wharf-cmd to access build results.
+// @description Please refer to the gRPC API for more endpoints.
+// @license.name MIT
+// @license.url https://github.com/iver-wharf/wharf-cmd/blob/master/LICENSE
+// @contact.name Iver wharf-cmd support
+// @contact.url https://github.com/iver-wharf/wharf-cmd/issues
+// @contact.email wharf@iver.se
+// @query.collection.format multi
 func serveHTTP(s *restServer, listener net.Listener) error {
 	r := gin.New()
 	applyGinHandlers(r)
 	applyCORSConfig(r)
 
-	g := r.Group("/api")
-	g.GET("", func(c *gin.Context) { c.JSON(200, gin.H{"message": "pong"}) })
+	r.GET("", pingHandler)
 
-	s.registerModules(g)
+	api := r.Group("/api")
+	api.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler, func(c *ginSwagger.Config) {
+		c.InstanceName = docs.SwaggerInfoworkerapi.InstanceName()
+	}))
+	artifactModule{s.artifactOpener}.register(api)
 	return r.RunListener(listener)
-}
-
-func (s *restServer) registerModules(r *gin.RouterGroup) {
-	modules := []module{
-		&artifactModule{s.artifactOpener},
-	}
-	for _, module := range modules {
-		module.register(r)
-	}
 }
 
 func applyGinHandlers(r *gin.Engine) {
@@ -57,4 +60,22 @@ func applyCORSConfig(r *gin.Engine) {
 	corsConfig := cors.DefaultConfig()
 	corsConfig.AllowAllOrigins = true
 	r.Use(cors.New(corsConfig))
+}
+
+// Ping is the response from a GET /api/ request.
+type Ping struct {
+	Message string `json:"message" example:"pong"`
+}
+
+// pingHandler godoc
+// @id ping
+// @summary Ping
+// @description Pong.
+// @description Added in v0.8.0.
+// @tags meta
+// @produce json
+// @success 200 {object} Ping
+// @router / [get]
+func pingHandler(c *gin.Context) {
+	c.JSON(200, Ping{Message: "pong"})
 }